Moves comments function into specific files
