+++ packages/prettier-plugin-java/src/printers/blocks-and-statements.js
@@ -4,6 +4,10 @@ const { line, softline, hardline } = require("prettier").doc.builders;
 const { group, indent, concat, join } = require("./prettier-builder");
 const { printTokenWithComments } = require("./comments/format-comments");
 const {
+  hasLeadingLineComments,
+  hasTrailingLineComments
+} = require("./comments/comments-utils");
+const {
   displaySemicolon,
   rejectAndConcat,
   rejectAndJoin,
@@ -12,9 +16,7 @@ const {
   rejectSeparators,
   putIntoBraces,
   isStatementEmptyStatement,
-  sortModifiers,
-  hasTrailingLineComments,
-  hasLeadingLineComments
+  sortModifiers
 } = require("./printer-utils");
 
 class BlocksAndStatementPrettierVisitor {
