+++ packages/prettier-plugin-java/src/printers/expressions.js
@@ -5,6 +5,9 @@ const { ifBreak, line, softline } = require("prettier").doc.builders;
 const { concat, group, indent } = require("./prettier-builder");
 const { printTokenWithComments } = require("./comments/format-comments");
 const {
+  handleCommentsBinaryExpression
+} = require("./comments/handle-comments");
+const {
   matchCategory,
   rejectAndJoin,
   rejectAndConcat,
@@ -15,8 +18,7 @@ const {
   putIntoBraces,
   separateTokensIntoGroups,
   isShiftOperator,
-  isUniqueMethodInvocation,
-  handleComments
+  isUniqueMethodInvocation
 } = require("./printer-utils");
 
 class ExpressionsPrettierVisitor {
@@ -159,7 +161,7 @@ class ExpressionsPrettierVisitor {
   }
 
   binaryExpression(ctx, params) {
-    handleComments(ctx);
+    handleCommentsBinaryExpression(ctx);
 
     const referenceType = this.mapVisit(ctx.referenceType);
     const expression = this.mapVisit(ctx.expression);
